chore: Move mock storage to tests
